requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.6.1
requests==2.21.0
toml==0.10.0
Werkzeug==0.14.1
